# docs: fix typo in latest note