@Tieske 2015-11-08
> Would you accept a PR with those globals cached?

Sure!